[02_MVC] Challanges 1 & 2
The two challanges involve adding click-and-next functionality to the
question text and a previous button that scrolls the question list
backward upon clicking.

By the way, we noted that all button texts are converted upper-case.
It turned out that this is performed by the app theme. Therefore, an
explicit style rule is added to prevent this.